Added AbstractVisitor.
