+++ src/database/src/Commands/Ast/ModelRewriteGetterSetterVisitor.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://hyperf.wiki
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
+ */
+namespace Hyperf\Database\Commands\Ast;
+
+class ModelRewriteGetterSetterVisitor extends AbstractVisitor
+{
+}
